Further clean up and dividing Reader and Writer
