+++ README.md
@@ -11,4 +11,4 @@ This project is meant for being used in the following way:
 ## How to use this project
 1. In the dk.cphbusiness.flightdemo.FlightReader class, you will find methods to read flight data from a json file into a list of DTO objects.
 2. To see an example of using streams to get info from the collection, look at the dk.cphbusiness.flightdemo.FlightReader.getFlightInfoDetails() method.
-3. In the test package, you will find a test class that tests the FlightReader class. Build on this to add more functionality to the FlightReader class. 
+3. In the test package, you will find a test class that tests the FlightReader and FlightWriter classes. Build on this to add more functionality to the FlightReader class.
