+++ src/main/java/dk/cphbusiness/utils/Utils.java
@@ -1,5 +1,11 @@
 package dk.cphbusiness.utils;
 
+import com.fasterxml.jackson.core.util.DefaultPrettyPrinter;
+import com.fasterxml.jackson.databind.DeserializationFeature;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
+import dk.cphbusiness.flightdemo.FlightReader;
+
 import java.io.IOException;
 import java.io.InputStream;
 import java.util.Properties;
@@ -9,19 +15,17 @@ import java.util.Properties;
  * Author: Thomas Hartmann
  */
 public class Utils {
-    public static void main(String[] args) {
-        System.out.println(getPropertyValue("db.name", "properties-from-pom.properties"));
+
+    public String getPropertyValue(String key) throws IOException {
+        Properties props = new Properties();
+        props.load(FlightReader.class.getClassLoader().getResourceAsStream("config.properties"));
+        return props.getProperty(key);
     }
-    public static String getPropertyValue(String propName, String ressourceName)  {
-        // REMEMBER TO BUILD WITH MAVEN FIRST. Read the property file if not deployed (else read system vars instead)
-        // Read from ressources/config.properties or from pom.xml depending on the ressourceName
-        try (InputStream is = Utils.class.getClassLoader().getResourceAsStream(ressourceName)) { //"config.properties" or "properties-from-pom.properties"
-            Properties prop = new Properties();
-            prop.load(is);
-            return prop.getProperty(propName);
-        } catch (IOException ex) {
-            ex.printStackTrace();
-        }
-        return null;
+    public static ObjectMapper getObjectMapper() {
+        ObjectMapper objectMapper = new ObjectMapper();
+        objectMapper.configure(DeserializationFeature.FAIL_ON_UNKNOWN_PROPERTIES, false);
+        objectMapper.registerModule(new JavaTimeModule());
+        objectMapper.writer(new DefaultPrettyPrinter());
+        return objectMapper;
     }
 }
